requirements.txt
openai-whisper==20250625
requests==2.32.0
watchdog==3.0.0
cryptography==42.0.4
python-dotenv==1.0.1
numpy<2.0
pyannote.audio==3.1.1
torch==2.7.1
torchaudio==2.7.1
librosa==0.10.1
soundfile==0.12.1
Flask==3.0.3
noisereduce==3.0.0
pydub==0.25.1
scipy>=1.9.0